fix: all tooltip tests and split out into relevant directory
