+++ packages/tooltip/__tests__/shared.js
@@ -1,13 +1,10 @@
 import Tooltip from "../tooltip";
 import TestRenderer from "react-test-renderer";
 import { ResponsiveContext } from "@times-components-native/responsive";
-import { delay } from "@times-components-native/test-utils";
 import { shallow } from "enzyme";
-import { TouchableOpacity, Animated } from "react-native";
 import React from "react";
 import { Text } from "react-native";
 import "./serializers-with-all-styles";
-import { Platform } from "react-native";
 
 export const withTabletContext = (WrappedComponent, isTablet = true) => (
   <ResponsiveContext.Provider
@@ -144,28 +141,6 @@ export default () => {
       expect(output).toMatchSnapshot();
     });
 
-    it("fades out when clicking the close button", async () => {
-      const onTooltipPresentedMock = jest.fn();
-
-      const output = shallow(
-        <Tooltip
-          content={<Text>foo</Text>}
-          onTooltipPresented={onTooltipPresentedMock}
-          type="testtype"
-          tooltips={["testtype"]}
-        >
-          bar
-        </Tooltip>,
-      );
-
-      const closeButton = output.find(TouchableOpacity);
-      closeButton.simulate("press");
-      await delay(250);
-      expect(output.find(Animated.View).get(0).props.style.opacity._value).toBe(
-        0,
-      );
-    });
-
     it("onTooltipPresented is called correctly on Viewport Enter", async () => {
       const onTooltipPresentedMock = jest.fn();
 
@@ -198,46 +173,6 @@ export default () => {
         </Tooltip>,
       );
       expect(onTooltipPresentedMock).toBeCalled();
-    });
-
-    it("onClose is called correctly", async () => {
-      const onTooltipPresentedMock = jest.fn();
-      const onCloseMock = jest.fn();
-
-      const testInstance = shallow(
-        <Tooltip
-          content={<Text>foo</Text>}
-          onClose={onCloseMock}
-          onTooltipPresented={onTooltipPresentedMock}
-          type="testtype"
-          tooltips={["testtype"]}
-        >
-          bar
-        </Tooltip>,
-      );
-
-      const closeButton = testInstance.find(TouchableOpacity);
-      closeButton.simulate("press");
-      expect(onCloseMock).toBeCalled();
-    });
-
-    it("hides tooltip close button on android only", () => {
-      Platform.OS = "android";
-
-      const onTooltipPresentedMock = jest.fn();
-
-      const output = TestRenderer.create(
-        <Tooltip
-          content={<Text>foo</Text>}
-          onTooltipPresented={onTooltipPresentedMock}
-          type="testtype"
-          tooltips={["testtype"]}
-          placement="top"
-        >
-          bar
-        </Tooltip>,
-      );
-      expect(output).toMatchSnapshot();
     });
   });
 };
